employees_web/requirements.txt
Django==2.2.4
djangorestframework==3.10.2
PyMySQL==0.8.1
mysqlclient==1.4.4
django-mptt==0.10.0
